Time: 464 ms (5.39%), Space: 61.3 MB (25.93%) - LeetHub

+++ 2695-find-score-of-an-array-after-marking-all-elements/2695-find-score-of-an-array-after-marking-all-elements.java
@@ -36,7 +36,7 @@ class Solution {
                 set.add(pair.getIndex() - 1);
                 set.add(pair.getIndex() + 1);
                 set.add(pair.getIndex());
-                ans += (long) pair.getVal();
+                ans += pair.getVal();
             }
         }
